Uporządkowanie klas w pakietach
