+++ kodilla-rps/src/main/java/com/kodilla/rps/figures/Figure.java
@@ -1,4 +1,4 @@
-package com.kodilla.rps;
+package com.kodilla.rps.figures;
 
 public interface Figure {
     int DRAW = 0;
