+++ kodilla-rps/src/main/java/com/kodilla/rps/figures/Paper.java
@@ -1,11 +1,11 @@
-package com.kodilla.rps;
+package com.kodilla.rps.figures;
 
-public class Rock implements Figure {
+public class Paper implements Figure {
 
     public int winWith(Figure figure) {
         if ( getClass().getSimpleName().equals(figure.getClass().getSimpleName()) ) {
             return DRAW;
-        } else if ( figure.getClass().getSimpleName().equals(Scissors.class.getSimpleName()) ) {
+        } else if ( figure.getClass().getSimpleName().equals(Rock.class.getSimpleName()) ) {
             return WIN;
         } else {
             return LOSS;
