+++ kodilla-rps/src/main/java/com/kodilla/rps/players/UserPlayer.java
@@ -1,8 +1,12 @@
-package com.kodilla.rps;
+package com.kodilla.rps.players;
 
 import com.kodilla.rps.exceptions.EndGameException;
 import com.kodilla.rps.exceptions.NewGameException;
 import com.kodilla.rps.exceptions.WrongNumberException;
+import com.kodilla.rps.figures.Figure;
+import com.kodilla.rps.figures.Paper;
+import com.kodilla.rps.figures.Rock;
+import com.kodilla.rps.figures.Scissors;
 
 import java.util.Scanner;
